Select V3 (#24)
* Can select and visually see rectangle

* realized i used the wrong editor.html so temp commit

* good editor

* Very basic selection works

* Changed logic

to fit closer to industry standard

* start of cancel/delete fix

* storage logic start

* Changed to use storage.py solutions

* improved storage useage

* Fixed my dumb merge mistake

* Select V2

Finalized
polygon select
freeform select

fixed bugs in
cop cut paste

Improved apply saving to fix a bug with undo

Added persistant clipboard (only when canvas is active, not sure why. but it works fine)

* small refactor to clean some lines

* can select the inside of objects really well

improved magic lasso logic a "bit" and renamed to magnetic as that is the industry name.

cleaned up a bit

---------

Co-authored-by: Teodor Log Bjørvik <teodorlb@uia.no>